pyproject.toml
[project]
name = "pubserv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ell-ai[all]>=0.0.15",
    "openai>=1.63.2",
    "pdf2image>=1.17.0",
    "polars>=1.22.0",
    "pydantic>=2.10.6",
    "tqdm>=4.67.1",
]
